Added Summary comments
