+++ Data-Structures/Trees/Trees/Tree.cs
@@ -14,11 +14,20 @@ namespace Trees
 
         }
 
+        /// <summary>
+        /// Tree constructor with generic Root node
+        /// </summary>
+        /// <param name="value"></param>
         public Tree(T value)
         {
             Root = new Node<T>(value);
         }
 
+        /// <summary>
+        /// PreOrder traversal of a BT
+        /// </summary>
+        /// <param name="root">Root of the tree</param>
+        /// <returns>List in PreOrder form of nodes in tree</returns>
         public List<T> PreOrder(Node<T> root)
         {
             List<T> result = new List<T>();
@@ -26,6 +35,11 @@ namespace Trees
             return result;
         }
 
+        /// <summary>
+        /// Recursive helper for PreOrder traversal
+        /// </summary>
+        /// <param name="result">List result to add to</param>
+        /// <param name="root">Current node in recursion, start at root</param>
         private void PreOrder(List<T> result, Node<T> root)
         {
             result.Add(root.Value);
@@ -39,6 +53,11 @@ namespace Trees
             }
         }
 
+        /// <summary>
+        /// InOrder traversal of a BT
+        /// </summary>
+        /// <param name="root">Root of the tree</param>
+        /// <returns>List in InOrder form of nodes in tree</returns>
         public List<T> InOrder(Node<T> root)
         {
             List<T> result = new List<T>();
@@ -46,6 +65,11 @@ namespace Trees
             return result;
         }
 
+        /// <summary>
+        /// Recursive helper for InOrder traversal
+        /// </summary>
+        /// <param name="result">List result to add to</param>
+        /// <param name="root">Current node in recursion, start at root</param>
         private void InOrder(List<T> result, Node<T> root)
         {
             if (root.LC != null)
@@ -59,6 +83,11 @@ namespace Trees
             }
         }
 
+        /// <summary>
+        /// PostOrder traversal of a BT
+        /// </summary>
+        /// <param name="root">Root of the tree</param>
+        /// <returns>List in PostOrder form of nodes in tree</returns>
         public List<T> PostOrder(Node<T> root)
         {
             List<T> result = new List<T>();
@@ -66,6 +95,11 @@ namespace Trees
             return result;
         }
 
+        /// <summary>
+        /// Recursive helper for PostOrder traversal
+        /// </summary>
+        /// <param name="result">List result to add to</param>
+        /// <param name="root">Current node in recursion, start at root</param>
         private void PostOrder(List<T> result, Node<T> root)
         {
             if (root.LC != null)
@@ -79,6 +113,11 @@ namespace Trees
             result.Add(root.Value);
         }
 
+        /// <summary>
+        /// Find the maximum value in a BT
+        /// </summary>
+        /// <param name="root">Current node in recursion, starts at tree's root</param>
+        /// <returns>Maximum value in tree</returns>
         public int FindMaximumValue(Node<int> root)
         {
             if (root == null)
